Created Global Grid system
prior to this update there was no "world map", the nsake tiles and apples were stored in seperate arrays, this would make it very difficult for the raycasting needed to let the AI's see. So now there is a global array storing everything which will make it easy for raycasting. Also the collisions have been changed.